* 'master' of gitlab.ayla.com.cn:ccpg-hotel/construction-android:
  1.pact修改
  1.场景阶段提交

+++ app/src/main/java/com/ayla/hotelsaas/mvp/present/SceneSettingPresenter.java
@@ -55,7 +55,7 @@ public class SceneSettingPresenter extends BasePresenter<SceneSettingView> {
                 });
     }
 
-    public void deleteScene(int ruleId) {
+    public void deleteScene(long ruleId) {
         RequestModel.getInstance()
                 .deleteRuleEngine(ruleId)
                 .subscribeOn(Schedulers.io())
